requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-community
email-validator
